+++ mods/CreateDragonPlus/CHANGELOG.md
@@ -3,4 +3,5 @@
 ### Fix
 * Fix tags fail to build if not all supported mods are installed.
 * Fix Create: Dreams n' Desires's Snow recipe does not work when combined with Garnished and Dragons Plus.
+* Fix Fog color of dye fluid not working correctly